Added argparse and arguments. Nesting functions.
Added the argparse module and named arguments for user interaction with
the script. The arguments assign variables to the path, filetypes,
searched for content, and replacement content. Nested the os.walk for
loop in a function named sleuth for cleanliness and usability. Started
generalizing objects so the script is flexible.