Working comments: remove old, add new. Fix gobble.
Removed old working comments, and added new working comments to notate
areas that I am working on. Maybe there is a better way to do this?

Fixed the exception block in the function gobble(). As of this commit
you can run the file in ipython using %run replacer.py [arguments] and
then execute the function main(pathname) which will populate the list
filestoedit. The gobble() function is the focus at this point, you can
manually create a tree from a file in filestoedit. See the comments
before gobble for specific info.