Wrote list comprehensions, grouping for gobble().
Added new code blocks to gobble().

Line 117 list comprehension applies the xpath arguments provided by the
user which are held in the container ekspath to the parsed tree. The
returned list 'oldcontent' is then flattened with another list
comprehension on line 119. This list contains elementtree objects for
each match returned from the tree.xpath search. There may be any number
of elements for each argument.

We want to insert our new content before each contiguous block of
elements. Then remove the old content. To group contiguous blocks of
elements together we use a for loop on line 123. This loop checks for
any blocks that evaluate to true for the etree.getprevious() function
If there are no contiguous elements we use the else statement on line
128 to index the original list of elements from line 119. The
targets_grouped list is an index of grouped elements, not the actual
elements. We may want to return the list of elements. We may also want
to break these blocks into separate funtions.